Split up base controller
1. Split up base controller
