+++ app/Http/Controllers/IndexController.php
@@ -42,116 +42,6 @@ class IndexController extends BaseController
         }
     }
 
-    public function summaries(Request $request)
-    {
-        $categories = null;
-        $years = null;
-        $this->nav_active = 'summaries';
-
-        $categories = Api::public()->get(
-            Config::get('web.config.api_uri_categories_summary'),
-            'IndexController@index'
-        );
-
-        $years = Api::public()->get(
-            Config::get('web.config.api_uri_years_summary'),
-            'IndexController@index'
-        );
-
-        if ($categories !== null && $years !== null) {
-            return view(
-                'summaries',
-                [
-                    'display_nav_options' => $this->display_nav_options,
-                    'nav_active' => $this->nav_active,
-                    'resource_name' => Config::get('web.config.api_resource_name'),
-                    'categories' => $categories,
-                    'years' => $years
-                ]
-            );
-        }
-    }
-
-    public function categoriesTco(Request $request)
-    {
-        $tco = null;
-        $this->nav_active = 'tco-summary';
-
-        $tco = Api::public()->get(
-            Config::get('web.config.api_uri_categories_tco'),
-            'IndexController@index'
-        );
-
-        if ($tco !== null) {
-            return view(
-                'tco-summary',
-                [
-                    'display_nav_options' => $this->display_nav_options,
-                    'nav_active' => $this->nav_active,
-                    'resource_name' => Config::get('web.config.api_resource_name'),
-                    'tco' => $tco
-                ]
-            );
-        }
-    }
-
-    public function subCategoriesSummary(Request $request, string $category_identifier)
-    {
-        $category = null;
-        $sub_categories = null;
-        $this->nav_active = 'summaries';
-
-        $category = Api::public()->get(
-            Config::get('web.config.api_uri_category') . '/' .
-            $category_identifier,
-            'IndexController@index'
-        );
-
-        $sub_categories = Api::public()->get(
-            Config::get('web.config.api_uri_categories_summary') .
-            '/' . $category_identifier . '/sub_categories',
-            'IndexController@index'
-        );
-
-        if ($category === null || $sub_categories !== null) {
-            return view(
-                'sub-categories-summary',
-                [
-                    'display_nav_options' => $this->display_nav_options,
-                    'nav_active' => $this->nav_active,
-                    'resource_name' => Config::get('web.config.api_resource_name'),
-                    'category' => $category,
-                    'sub_categories' => $sub_categories
-                ]
-            );
-        }
-    }
-
-    public function addExpense(Request $request)
-    {
-        $this->nav_active = 'add-expense';
-
-        $sub_categories = Api::public()->get(
-            Config::get('web.config.api_uri_categories') .
-            '/' . Config::get('web.config.api_category_id_essentials') .
-            '/sub_categories',
-            'IndexController@index'
-        );
-
-        return view(
-            'add-expense',
-            [
-                'display_nav_options' => $this->display_nav_options,
-                'nav_active' => $this->nav_active,
-                'resource_name' => Config::get('web.config.api_resource_name'),
-                'category_id_essentials' => Config::get('web.config.api_category_id_essentials'),
-                'category_id_non_essentials' => Config::get('web.config.api_category_id_non_essentials'),
-                'category_id_hobbies_and_interests' => Config::get('web.config.api_category_id_hobbies_and_interests'),
-                'sub_categories' => $sub_categories
-            ]
-        );
-    }
-
     public function subCategories(Request $request, string $category_identifier)
     {
         $sub_categories = Api::public()->get(
@@ -161,324 +51,6 @@ class IndexController extends BaseController
         );
 
         return response()->json($sub_categories, 200);
-    }
-
-    public function processAddExpense(Request $request)
-    {
-        $item = null;
-        $item_category = null;
-        $item_sub_category = null;
-
-        $client = new Client([
-            'base_uri' => Config::get('web.config.api_base_url'),
-            'headers' => [
-                'Authorization' => 'Bearer ' . $request->session()->get('bearer'),
-                'Accept' => 'application/json',
-                'Content-Type' => 'application/json',
-            ],
-        ]);
-
-        try {
-            $response = $client->post(
-                Config::get('web.config.api_uri_items'),
-                [
-                    \GuzzleHttp\RequestOptions::JSON => [
-                        'description' => $request->input('description'),
-                        'effective_date' => $request->input('effective_date'),
-                        'total' => $request->input('total'),
-                        'percentage' => $request->input('allocation')
-                    ]
-                ]
-            );
-
-            if ($response->getStatusCode() === 201) {
-                $item = json_decode($response->getBody(), true);
-            } else {
-                // Check for 422 (validation) and then display below for general errors
-                $request->session()->flash('status', 'expense-not-added-item');
-                return redirect()->action('IndexController@recent');
-            }
-        } catch (ClientException $e) {
-            $request->session()->flash('status', 'api-error');
-            $request->session()->flash('status-line', __LINE__);
-            return redirect()->action('IndexController@recent');
-        }
-
-        try {
-            $response = $client->post(
-                Config::get('web.config.api_uri_items') . '/' . $item['id'] . '/category',
-                [
-                    \GuzzleHttp\RequestOptions::JSON => [
-                        'category_id' => $request->input('category_id')
-                    ]
-                ]
-            );
-
-            if ($response->getStatusCode() === 201) {
-                $item_category = json_decode($response->getBody(), true);
-            } else {
-                // Check for 422 (validation) and then display below for general errors
-                $request->session()->flash('status', 'expense-not-added-item-category');
-                return redirect()->action('IndexController@recent');
-            }
-        } catch (ClientException $e) {
-            $request->session()->flash('status', 'api-error');
-            $request->session()->flash('status-line', __LINE__);
-            return redirect()->action('IndexController@recent');
-        }
-
-        try {
-            $response = $client->post(
-                Config::get('web.config.api_uri_items') . '/' .
-                $item['id'] . '/category/' . $item_category['id'] . '/sub_category',
-                [
-                    \GuzzleHttp\RequestOptions::JSON => [
-                        'sub_category_id' => $request->input('sub_category_id')
-                    ]
-                ]
-            );
-
-            if ($response->getStatusCode() === 201) {
-                $item_sub_category = json_decode($response->getBody(), true);
-            } else {
-                // Check for 422 (validation) and then display below for general errors
-                $request->session()->flash('status', 'expense-not-added-item-sub-category');
-                return redirect()->action('IndexController@recent');
-            }
-        } catch (ClientException $e) {
-            $request->session()->flash('status', 'api-error');
-            $request->session()->flash('status-line', __LINE__);
-            return redirect()->action('IndexController@recent');
-        }
-
-        if ($item !== null && $item_category !== null && $item_sub_category !== null) {
-            $request->session()->flash('status', 'expense-added');
-            return redirect()->action('IndexController@recent');
-        } else {
-            $request->session()->flash('status', 'expense-not-added');
-            return redirect()->action('IndexController@recent');
-        }
-    }
-
-    public function expense(Request $request, string $expense_identifier)
-    {
-        $expense = null;
-        $category = null;
-        $sub_category = null;
-        $status = null;
-        $this->nav_active = 'recent';
-
-        $expense = Api::public()->get(
-            Config::get('web.config.api_uri_items') . '/' .
-            $expense_identifier,
-            'IndexController@index'
-        );
-
-        $category = Api::public()->get(
-            Config::get('web.config.api_uri_items') . '/' .
-            $expense_identifier . '/category',
-            'IndexController@index'
-        );
-
-        if ($category === null) {
-            $status = 'category-not-assigned';
-        }
-
-        if ($category !== null) {
-            $sub_category = Api::public()->get(
-                Config::get('web.config.api_uri_items') . '/' .
-                $expense_identifier . '/category/' . $category['id'] .
-                '/sub_category',
-                'IndexController@index'
-            );
-
-            if ($sub_category === null) {
-                $status = 'sub-category-not-assigned';
-            }
-        } else {
-            $sub_category = null;
-        }
-
-
-        if ($expense !== null) {
-            return view(
-                'expense',
-                [
-                    'display_nav_options' => $this->display_nav_options,
-                    'nav_active' => $this->nav_active,
-                    'expense' => $expense,
-                    'category' => $category,
-                    'sub_category' => $sub_category,
-                    'status' => $status
-                ]
-            );
-        }
-    }
-
-    public function deleteExpense(Request $request, string $expense_identifier)
-    {
-        $expense = null;
-        $category = null;
-        $sub_category = null;
-
-        $expense_identifier_id = null;
-        $expense_category_identifier_id = null;
-        $expense_sub_category_identifier_id = null;
-
-        $this->nav_active = 'recent';
-
-        $expense = Api::public()->get(
-            Config::get('web.config.api_uri_items') . '/' .
-            $expense_identifier,
-            'IndexController@index'
-        );
-
-        if ($expense !== null) {
-            $expense_identifier_id = $expense['id'];
-        }
-
-        $category = Api::public()->get(
-            Config::get('web.config.api_uri_items') . '/' .
-            $expense_identifier . '/category',
-            'IndexController@index'
-        );
-
-        if ($category !== null) {
-            $expense_category_identifier_id = $category['id'];
-        }
-
-        if ($category !== null) {
-            $sub_category = Api::public()->get(
-                Config::get('web.config.api_uri_items') . '/' .
-                $expense_identifier . '/category/' . $category['id'] .
-                '/sub_category',
-                'IndexController@index'
-            );
-
-            if ($sub_category !== null) {
-                $expense_sub_category_identifier_id = $sub_category['id'];
-            }
-        }
-
-        if ($expense !== null) {
-            return view(
-                'delete-expense',
-                [
-                    'display_nav_options' => $this->display_nav_options,
-                    'nav_active' => $this->nav_active,
-                    'expense' => $expense,
-                    'category' => $category,
-                    'sub_category' => $sub_category,
-                    'expense_identifier_id' => $expense_identifier_id,
-                    'expense_category_identifier_id' => $expense_category_identifier_id,
-                    'expense_sub_category_identifier_id' => $expense_sub_category_identifier_id
-                ]
-            );
-        }
-    }
-
-    public function processDeleteExpense(Request $request)
-    {
-        $client = new Client([
-            'base_uri' => Config::get('web.config.api_base_url'),
-            'headers' => [
-                'Authorization' => 'Bearer ' . $request->session()->get('bearer'),
-                'Accept' => 'application/json',
-                'Content-Type' => 'application/json',
-            ],
-        ]);
-
-        $deleted_sub_category = false;
-        $deleted_category = false;
-        $deleted_expense = false;
-
-        $expense_identifier = $request->input('expense_identifier_id');
-        $expense_category_identifier = $request->input('expense_category_identifier_id');
-        $expense_sub_category_identifier = $request->input('expense_sub_category_identifier_id');
-
-        if ($expense_sub_category_identifier !== null) {
-            try {
-                $response = $client->delete(
-                    Config::get('web.config.api_uri_items') . '/' .
-                    $expense_identifier . '/category/' . $expense_category_identifier .
-                    '/sub_category/' . $expense_sub_category_identifier
-                );
-
-                if ($response->getStatusCode() === 204) {
-                    $deleted_sub_category = true;
-                }
-            } catch (ClientException $e) {
-                // Ignore for now until logging added
-            }
-        } else {
-            $deleted_sub_category = true;
-        }
-
-        if ($deleted_sub_category === true && $expense_category_identifier !== null) {
-            try {
-                $response = $client->delete(
-                    Config::get('web.config.api_uri_items') . '/' .
-                    $expense_identifier . '/category/' . $expense_category_identifier
-                );
-
-                if ($response->getStatusCode() === 204) {
-                    $deleted_category = true;
-                }
-            } catch (ClientException $e) {
-                // Ignore for now until logging added
-            }
-        } else {
-            $deleted_category = true;
-        }
-
-        if ($deleted_category === true && $expense_identifier !== null) {
-            try {
-                $response = $client->delete(
-                    Config::get('web.config.api_uri_items') . '/' . $expense_identifier
-                );
-
-                if ($response->getStatusCode() === 204) {
-                    $deleted_expense = true;
-                }
-            } catch (ClientException $e) {
-                // Ignore for now until logging added
-            }
-        } else {
-            $deleted_expense = true;
-        }
-
-        if ($deleted_expense === true && $deleted_category === true && $deleted_sub_category === true) {
-            $request->session()->flash('status', 'expense-deleted');
-            return redirect()->action('IndexController@recent');
-        } else {
-            $request->session()->flash('status', 'expense-not-deleted');
-            return redirect()->action('IndexController@expense', ['expense_identifier' => $expense_identifier]);
-        }
-    }
-
-    public function monthsSummary(Request $request, string $year_identifier)
-    {
-        $months = null;
-        $this->nav_active = 'summaries';
-
-        $months = Api::public()->get(
-            Config::get('web.config.api_uri_years_summary') .
-            '/' . $year_identifier . '/months',
-            'IndexController@index'
-        );
-
-        if ($months !== null) {
-            return view(
-                'months-summary',
-                [
-                    'display_nav_options' => $this->display_nav_options,
-                    'nav_active' => $this->nav_active,
-                    'resource_name' => Config::get('web.config.api_resource_name'),
-                    'months' => $months,
-                    'year' => $year_identifier
-                ]
-            );
-        }
     }
 
     public function versionHistory()
@@ -494,81 +66,5 @@ class IndexController extends BaseController
                 'version_history' => $html
             ]
         );
-    }
-
-    public function expenses(Request $request)
-    {
-        $expenses = null;
-
-        $allowed = ['year', 'month', 'category', 'sub_category'];
-        $request_parameters = [];
-        $filtering = [];
-        $parameters = $request->all();
-
-        foreach ($parameters as $parameter => $value) {
-            if (
-                in_array($parameter, $allowed) === true &&
-                $value !== null
-            ) {
-                switch ($parameter) {
-                    case 'year':
-                        $filtering[] = 'Year: ' . $value;
-                        break;
-                    case 'month':
-                        $filtering[] = 'Month: ' . date("F", mktime(0, 0, 0, $value, 10));
-                        break;
-                    case 'category':
-
-                        $category = Api::public()->get(
-                            Config::get('web.config.api_uri_category') .
-                            '/' . $value,
-                            'IndexController@index'
-                        );
-
-                        if ($category !== null) {
-                            $filtering[] .= $category['name'];
-                        }
-                        break;
-                    case 'sub_category':
-                        try {
-                            $response = $client->get(
-                                Config::get('web.config.api_uri_category') .
-                                '/' . $request_parameters['category'] . '/sub_categories/' . $value
-                            );
-
-                            if ($response->getStatusCode() === 200) {
-                                $filtering[] .= 'Sub category: ' . json_decode($response->getBody(), true)['name'];
-                            }
-                        } catch (ClientException $e) {
-                            // Do nothing for now
-                        }
-                        break;
-                    default:
-                        break;
-                }
-
-                $request_parameters[$parameter] = $value;
-            }
-        }
-
-        $uri = Config::get('web.config.api_uri_items') . '?limit=5';
-        foreach ($request_parameters as $parameter => $value) {
-            $uri .= '&' . $parameter . '=' . $value;
-        }
-
-        $expenses = Api::public()->get($uri,'IndexController@index');
-
-        if ($expenses !== null) {
-            return view(
-                'expenses',
-                [
-                    'display_nav_options' => $this->display_nav_options,
-                    'nav_active' => $this->nav_active,
-                    'resource_name' => Config::get('web.config.api_resource_name'),
-                    'expenses' => $expenses,
-                    'filtering' => implode(', ', $filtering)
-                ]
-            );
-        }
     }
 }
