+++ resources/views/delete-expense.blade.php
@@ -6,7 +6,7 @@
 
         <p class="lead">Are you sure you want to delete this expense?</p>
 
-        <p><a href="{{ action('IndexController@expense', ['expense_identifier' => $expense['id']]) }}" class="btn btn-sm btn-outline-info">Return to expense</a></p>
+        <p><a href="{{ action('ExpenseController@expense', ['expense_identifier' => $expense['id']]) }}" class="btn btn-sm btn-outline-info">Return to expense</a></p>
 
         <dl class="row">
             <dt class="col-4">Description:</dt>
@@ -36,7 +36,7 @@
         </dl>
     </div>
     <div class="col-12 text-left mt-2">
-        <form method="post" action="{{ action('IndexController@processDeleteExpense') }}">
+        <form method="post" action="{{ action('ProcessController@processDeleteExpense') }}">
             <div class="form-group">
                 {{ csrf_field() }}
                 <input type="hidden" name="expense_identifier_id" value="{{ $expense_identifier_id }}" />
