+++ resources/views/expenses.blade.php
@@ -7,7 +7,7 @@
         <p class="lead">The last 5 expenses that have been entered for {{ $resource_name }},
             list filtering, {{ $filtering }}</p>
 
-        <p><a href="{{ action('IndexController@summaries') }}" class="btn btn-sm btn-outline-info">Return to summaries</a></p>
+        <p><a href="{{ action('SummaryController@summaries') }}" class="btn btn-sm btn-outline-info">Return to summaries</a></p>
 
         <table class="table table-sm">
             <caption>Filtered recent expenses added for {{ $resource_name }}</caption>
@@ -22,7 +22,7 @@
             <tbody>
                 @foreach ($expenses as $expense)
                 <tr class="d-table-row d-sm-none">
-                    <td colspan="4"><strong><a href="{{ action('IndexController@expense', ['expense_identifier' => $expense['id']]) }}" class="text-info">{{ $expense['description'] }}</a></strong></td>
+                    <td colspan="4"><strong><a href="{{ action('ExpenseController@expense', ['expense_identifier' => $expense['id']]) }}" class="text-info">{{ $expense['description'] }}</a></strong></td>
                 </tr>
                 <tr class="d-table-row d-sm-none">
                     <td colspan="2">{{ date('dS F Y', strtotime($expense['effective_date'])) }}</td>
@@ -30,7 +30,7 @@
                     <td>&nbsp;</td>
                 </tr>
                 <tr class="d-none d-sm-table-row">
-                    <td><strong><a href="{{ action('IndexController@expense', ['expense_identifier' => $expense['id']]) }}" class="text-info">{{ $expense['description'] }}</a></strong></td>
+                    <td><strong><a href="{{ action('ExpenseController@expense', ['expense_identifier' => $expense['id']]) }}" class="text-info">{{ $expense['description'] }}</a></strong></td>
                     <td>{{ date('dS M', strtotime($expense['effective_date'])) }}</td>
                     <td>&pound;{{ $expense['actualised_total'] }}</td>
                     <td>&nbsp;</td>
