+++ resources/views/sub-categories-summary.blade.php
@@ -12,7 +12,7 @@
             <dd class="col-9">{{ $category['description'] }}</dd>
         </dl>
 
-        <p><a href="{{ action('IndexController@summaries') }}" class="btn btn-sm btn-outline-info">Return to summaries</a></p>
+        <p><a href="{{ action('SummaryController@summaries') }}" class="btn btn-sm btn-outline-info">Return to summaries</a></p>
 
         <table class="table table-sm">
             <caption>Expenses summed by {{ $category['name'] }} category.</caption>
@@ -28,7 +28,7 @@
                 <tr>
                     <td><strong>{{ $sub_category['name'] }}</strong></td>
                     <td>&pound;{{ $sub_category['total'] }}</td>
-                    <td><a href="{{ action('IndexController@expenses', ['category' => $category['id'], 'sub_category' => $sub_category['id']]) }}" class="text-info"><i class="fas fa-list"></i></a></td>
+                    <td><a href="{{ action('ExpenseController@expenses', ['category' => $category['id'], 'sub_category' => $sub_category['id']]) }}" class="text-info"><i class="fas fa-list"></i></a></td>
                 </tr>
                 @endforeach
             </tbody>
