+++ resources/views/summaries.blade.php
@@ -21,9 +21,9 @@
             <tbody>
                 @foreach ($categories as $category)
                 <tr>
-                    <td><strong><a href="{{ action('IndexController@subCategoriesSummary', ['category_identifier' => $category['id']]) }}" class="text-info">{{ $category['name'] }}</a></strong></td>
+                    <td><strong><a href="{{ action('SummaryController@subCategoriesSummary', ['category_identifier' => $category['id']]) }}" class="text-info">{{ $category['name'] }}</a></strong></td>
                     <td>&pound;{{ $category['total'] }}</td>
-                    <td><a href="{{ action('IndexController@expenses', ['category' => $category['id']]) }}" class="text-info"><i class="fas fa-list"></i></a></td>
+                    <td><a href="{{ action('ExpenseController@expenses', ['category' => $category['id']]) }}" class="text-info"><i class="fas fa-list"></i></a></td>
                 </tr>
                 @endforeach
             </tbody>
@@ -46,9 +46,9 @@
             <tbody>
             @foreach ($years as $year)
                 <tr>
-                    <td><strong><a href="{{ action('IndexController@monthsSummary', ['year_identifier' => $year['id']]) }}" class="text-info">{{ $year['year'] }}</a></strong></td>
+                    <td><strong><a href="{{ action('SummaryController@monthsSummary', ['year_identifier' => $year['id']]) }}" class="text-info">{{ $year['year'] }}</a></strong></td>
                     <td>&pound;{{ $year['total'] }}</td>
-                    <td><a href="{{ action('IndexController@expenses', ['year' => $year['id']]) }}" class="text-info"><i class="fas fa-list"></i></a></td>
+                    <td><a href="{{ action('ExpenseController@expenses', ['year' => $year['id']]) }}" class="text-info"><i class="fas fa-list"></i></a></td>
                 </tr>
             @endforeach
             </tbody>
