+++ routes/web.php
@@ -13,17 +13,17 @@ Route::group(
     ],
     function () {
         Route::get('/recent', 'IndexController@recent');
-        Route::get('/summaries', 'IndexController@summaries');
-        Route::get('/sub-categories-summary/{category_identifier}', 'IndexController@subCategoriesSummary');
-        Route::get('/tco-summary', 'IndexController@categoriesTco');
-        Route::get('/months-summary/{year_identifier}', 'IndexController@monthsSummary');
-        Route::get('/add-expense', 'IndexController@addExpense');
+        Route::get('/summaries', 'SummaryController@summaries');
+        Route::get('/sub-categories-summary/{category_identifier}', 'SummaryController@subCategoriesSummary');
+        Route::get('/tco-summary', 'SummaryController@categoriesTco');
+        Route::get('/months-summary/{year_identifier}', 'SummaryController@monthsSummary');
+        Route::get('/add-expense', 'ExpenseController@addExpense');
         Route::get('/sub-categories/{category_identifier}', 'IndexController@subCategories');
-        Route::get('/expense/{expense_identifier}', 'IndexController@expense');
-        Route::post('/add-expense', 'IndexController@processAddExpense');
-        Route::get('/delete-expense/{expense_identifier}', 'IndexController@deleteExpense');
-        Route::post('/delete-expense', 'IndexController@processDeleteExpense');
+        Route::get('/expense/{expense_identifier}', 'ExpenseController@expense');
+        Route::post('/add-expense', 'ProcessController@processAddExpense');
+        Route::get('/delete-expense/{expense_identifier}', 'ExpenseController@deleteExpense');
+        Route::post('/delete-expense', 'ProcessController@processDeleteExpense');
         Route::get('/version-history', 'IndexController@versionHistory');
-        Route::get('/expenses', 'IndexController@expenses');
+        Route::get('/expenses', 'ExpenseController@expenses');
     }
 );
